app/requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.1.2
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
daphne==3.0.2
Django==3.2.9
djangochannelsrestframework==0.3.0
djangorestframework==3.12.4
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
kombu==5.1.0
msgpack==1.0.2
prompt-toolkit==3.0.20
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pillow==7.1.1
pyOpenSSL==20.0.1
pytz==2021.1
psycopg2-binary==2.9.1
redis==3.5.3
rest-framework-generic-relations==2.1.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0
django-silk
pytest
PyJWT==2.0.1
djangorestframework-simplejwt==4.8.0
djoser
drf-spectacular
django_filter
django-channels-jwt-auth-middleware
drf-yasg==1.20.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
